% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.davg.oncols.r
\name{apply.davg.oncols}
\alias{apply.davg.oncols}
\title{Apply daily average on columns of dataframe}
\usage{
apply.davg.oncols(
  mod.wo,
  daystep = 1,
  ignoreQzeros = F,
  calc.max = F,
  calc.min = F
)
}
\arguments{
\item{mod.wo}{character; A data frame with JDAY (numeric) as the first column in sub-daily timesteps}

\item{daystep}{numeric; timestep (in days) to calculate average over}

\item{ignoreQzeros}{logical; ignores zero flows in daily averaging !!!}

\item{calc.max}{logical; if true, calculate the daily max, not the daily average!}

\item{calc.min}{logical; if true, calculate the daily min, not the daily average!}
}
\value{
\item{davg.out}{a dataframe with daily data instead of sub-daily}
}
\description{
Apply daily average on columns of dataframe
}
\examples{
apply.davg.oncols()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{average}
\keyword{balance}
\keyword{daily}
\keyword{water}
