% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmergenceTiming.r
\name{calcEmergenceTiming}
\alias{calcEmergenceTiming}
\title{Calculate the estimated egg emergence timing based on cumulative degree-days of water temperature time-series data}
\usage{
calcEmergenceTiming(
  tout,
  atu.day = 263,
  davg = T,
  interpMissing = T,
  hatchValue = 1750
)
}
\arguments{
\item{tout}{dataframe; tout is a dataframe with the first column as a numeric Julian day}

\item{atu.day}{numeric; the day in which to start the emergence calculation, default = 263}

\item{davg}{logical; calculate based on daily average, default = TRUE}

\item{interpMissing}{logical, interpolate missing values? Default = TRUE}

\item{hatchValue}{numeric, the value in which eggs are assumed to hatch. Defaults to 1750}
}
\value{
\item{atu.d}{a dataframe with numeric JDAY dates}
\item{atu}{a dataframe with month and day units for the date}
}
\description{
Calculate the estimated egg emergence timing based on cumulative degree-days of water temperature time-series data
}
\examples{
calcEmergenceTiming()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{calculate}
\keyword{emergence}
\keyword{temperature}
\keyword{timing}
\keyword{water}
