% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errs.r
\name{errs}
\alias{errs}
\title{Calculate fit statistics}
\usage{
errs(out)
}
\arguments{
\item{out}{dataframe with measured data in first column, simulated in second column. No missing values allowed.}
}
\value{
\item{a dataframe with the number of observations (N), root mean squared error (RMSE), nash-sutcliffe error (NS), mean error (ME), mean absolute error (MAE), Kendall tau correlation coefficient (COR), SMAPE=SMAPE}
}
\description{
Function to compute the fit statistics of model and measured data v0.1 2020-03-02
}
\examples{
errs()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{balance}
\keyword{error}
\keyword{fit}
\keyword{water}
