% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dam.outflow.mod.output.r
\name{get.dam.outflow.mod.output}
\alias{get.dam.outflow.mod.output}
\title{This function will extract simulated CE-QUAL-W2 outflows from a dam}
\usage{
get.dam.outflow.mod.output(
  qwo.textfile = NA,
  seg = NA,
  path = "/W2_mods/Detroit_Lake_2010_00-01",
  metric = F,
  two.textfile = NA,
  str.txt = NA,
  withdrawals = 0,
  const = "ELWS"
)
}
\arguments{
\item{qwo.textfile}{character; name of the model outflow file to read}

\item{two.textfile}{character; name of the model outflow temperature file to read}

\item{str.txt}{character; prefix of the name of the model str file to read}

\item{withdrawals}{numeric; how many withdrawals; defaulst to 0}

\item{const}{character; constituent to read - defaults to 'ELWS'}
}
\value{
a data.frame of the model output
}
\description{
This function will extract simulated CE-QUAL-W2 outflows from a dam
}
\examples{
get.dam.outflow.mod.output()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{model}
\keyword{output}
\keyword{post-processing}
