% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.w2.inflows.r
\name{get.w2.inflows}
\alias{get.w2.inflows}
\title{This function will extract the inflows to a CE-QUAL-W2 model}
\usage{
get.w2.inflows(
  path = NA,
  q.rows = list(QIN = 1:4, QTR = 1:2, QDT = 1),
  d.avg = T
)
}
\arguments{
\item{path}{character; model directory to read from}

\item{q.rows}{list; rows of the control file to read defaults to q.rows=list(QIN=1:4,QTR=1:2,QDT=1)}

\item{d.avg}{logical; daily averaging? Defaults to TRUE}
}
\value{
a data.frame with the model inflows
}
\description{
This function will extract the inflows to a CE-QUAL-W2 model
}
\examples{
get.w2.inflows()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{inflows}
\keyword{model}
\keyword{output}
\keyword{post-processing}
\keyword{read}
\keyword{sum}
