% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getW2_41out.r
\name{getW2_41out}
\alias{getW2_41out}
\title{This function will extract simulated CE-QUAL-W2 lake elevation, outflows, inflows, and outflow temperature}
\usage{
getW2_41out(
  mod.path = w2Dirs[w2ioi],
  atu.day = atu.day,
  tcrit = NULL,
  modOutDetails = NA,
  sumOutTable = T,
  davg = c("avg", "max", "min")
)
}
\arguments{
\item{mod.path}{character; Defaults to reading in the wl.opt file, but can specify a "*.tsr" output from W2}

\item{atu.day}{numeric; the days in which the accumulated thermal units will begin. Typically c(244, 280, 310)}

\item{tcrit}{Optional dataframe with critical fish temperatures and days}

\item{modOutDetails}{Dataframe with columns: Dir, SegmentOutput, and wb columns}

\item{sumOutTable}{Logical; include the summary table output defaults to TRUE}

\item{davg}{character vector, typically c('avg','max','min')}
}
\value{

}
\description{
This function will extract simulated CE-QUAL-W2 lake elevation, outflows, inflows, and outflow temperature
}
\examples{
getW2_41out()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{model}
\keyword{output}
\keyword{post-processing}
