% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.by.jday.r
\name{merge.interp.24hr}
\alias{merge.interp.24hr}
\title{Merge 2 dataframes by JDAY with hourly data}
\usage{
\method{merge}{interp.24hr}(x1, x2, return.nas = T)
}
\arguments{
\item{x1}{dataframe; with JDAY column and another column with the data}

\item{x2}{dataframe; with JDAY column and another column with the data}

\item{return.nas}{logical; Include missing values?}
}
\value{
a dataframe with JDAY as the first column
}
\description{
Merge 2 dataframes by JDAY with hourly data
}
\examples{
merge.interp.24hr()
}
\author{
Norman Buccola
}
\keyword{JDAY}
\keyword{by}
\keyword{dataframes}
\keyword{merge}
