% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.by.jday.r
\name{merge.interp}
\alias{merge.interp}
\title{Merge 2 dataframes by JDAY}
\usage{
\method{merge}{interp}(x1, x2, return.nas = T)
}
\arguments{
\item{return.nas}{logical; Include missing values?}

\item{df1}{dataframe; with JDAY column and another column with the data}

\item{df2}{dataframe; with JDAY column and another column with the data}
}
\value{
a dataframe with JDAY as the first column
}
\description{
Merge 2 dataframes by JDAY
}
\examples{
merge.by.jday()
}
\author{
Norman Buccola
}
\keyword{2}
\keyword{JDAY}
\keyword{by}
\keyword{dataframes}
\keyword{merge}
