% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyInitWSELVbth.r
\name{modifyInitWSELVbth}
\alias{modifyInitWSELVbth}
\title{Modify initial water surface elevation in bathymetry file}
\usage{
modifyInitWSELVbth(path, newInitWSELV = NA, wb)
}
\arguments{
\item{path}{Define the model path in a character string}

\item{newInitWSELV}{numeric; New initial WSELV to be in meters - note: If not NA, the initial water level in the bathymetry file will be re-written to this value!}

\item{wb}{numeric CE-QUAL-W2 waterbody to look at}
}
\description{
Function to re-write the initial water surface within the bathymetry file of a CE-QUAL-W2 model v0.3 2020-03-02
}
\examples{
modifyInitWSELVbth()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{balance}
\keyword{water}
