% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyW2con.r
\name{modifyW2con}
\alias{modifyW2con}
\title{This function will modify the control file for a CE-QUAL-W2 model (w2_con.npt)}
\usage{
modifyW2con(path, qdt = "OFF", seg, branch)
}
\arguments{
\item{path}{character; model directory to read from}

\item{seg}{numeric; which segment to extract from the model?}

\item{branch}{numeric; which branch to extract from the model?}
}
\value{
re-writes a new w2_con.npt file
}
\description{
This function will modify the control file for a CE-QUAL-W2 model (w2_con.npt)
}
\examples{
modifyW2con()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{model}
\keyword{read}
\keyword{w2_con.npt}
