% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.avg.r
\name{moving.avg.oncols}
\alias{moving.avg.oncols}
\title{Calculate moving averages on columns}
\usage{
moving.avg.oncols(x, nwindow)
}
\arguments{
\item{x}{dataframe; with JDAY column and another column with the data}

\item{nwindow}{numeric; averaging window}
}
\value{
a dataframe with daily averages
}
\description{
Calculate moving averages on columns
}
\examples{
moving.avg.oncols()
}
\author{
Norman Buccola
}
\keyword{average}
\keyword{columns}
\keyword{moving}
\keyword{on}
