% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.avg.r
\name{moving.avg}
\alias{moving.avg}
\title{Calculate moving averages}
\usage{
moving.avg(x, nwindow)
}
\arguments{
\item{x}{dataframe; with JDAY column and another column with the data}

\item{nwindow}{numeric; averaging window}
}
\value{
a dataframe with daily averages
}
\description{
Calculate moving averages
}
\examples{
moving.avg()
}
\author{
Norman Buccola
}
\keyword{average}
\keyword{moving}
