% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readW2ConInOut.r
\name{readW2ConInOut}
\alias{readW2ConInOut}
\title{Wrapper function for waterBalance function}
\usage{
readW2ConInOut(path, wd, seg, wb, RESSIMCode, write.files)
}
\arguments{
\item{path, }{path to model}

\item{wd, }{Generic path to find functions}

\item{seg, }{#Segment to look up in w2_con.npt}

\item{wb, }{#Water body to look up in w2_con.npt}

\item{RESSIMCode, }{#RES-SIM name for boundary conditions}

\item{write.files}{#T/F on whether or not to write file}
}
\value{
\item{watbal}{output from the waterBalance function}
}
\description{
This function will grab inputs from CE-QUAL-W2 model (w2_con.npt file), lookup filenames, and then use them as input to the waterBalance function within the w2r package.
}
\examples{
readW2ConInOut()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{QDT}
\keyword{balance}
\keyword{water}
