% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readW2Selective.r
\name{readW2Selective}
\alias{readW2Selective}
\title{Read in the w2_selective.npt file for the CE-QUAL-W2 model}
\usage{
readW2Selective(path = path)
}
\arguments{
\item{path}{character; path to model}
}
\value{
\item{out}{a list with the number of outlets and priorities of each group}
}
\description{
Read in the w2_selective.npt file for the CE-QUAL-W2 model
}
\examples{
readW2ConInOut()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{read}
\keyword{w2_selective.npt}
