% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreQDT.r
\name{restoreQDT}
\alias{restoreQDT}
\title{Restore Previous QDT File}
\usage{
restoreQDT(path)
}
\arguments{
\item{path}{The model directory.}
}
\value{
TRUE if backup was restored.
}
\description{
Restore the QDT from a backup, if available.
}
