% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterBalanceGeneric.r
\name{waterBalance}
\alias{waterBalance}
\title{Water Balance Utility for CE-QUAL-W2 model}
\usage{
waterBalance(
  opt.txt = NA,
  seg = NA,
  wb = NA,
  new.npt.filename = NA,
  rule.curve.filename = NA,
  daystep = 1,
  path = "/*",
  meas.elvs = "***.npt",
  elvVolCrvFl = NA,
  write.files = T,
  save.plot = F,
  version = 4,
  append.filename = NULL,
  maxElvDif_m = 0.75
)
}
\arguments{
\item{opt.txt}{Defaults to reading in the wl.opt file, but can specify a "*.tsr" output from W2}

\item{seg}{numeric; which segment to get wselv from W2 output}

\item{wb}{numeric; which water body to get from W2 model output}

\item{new.npt.filename}{string new balance flow file to write}

\item{rule.curve.filename}{string filename for the optional Lake rule curve (csv format); if none then NA}

\item{daystep}{numeric number of days of averaging window; default to use the smallest timestep in model output}

\item{meas.elvs}{string filename for the measured elevations (csv format)}

\item{elvVolCrvFl}{string #Elevation - Volume Curve filename}

\item{write.files}{logical Write a new distributed tributary inflow file? TRUE or FALSE}

\item{save.plot}{logical Save a plot of the output? TRUE or FALSE}

\item{version}{numeric CE-QUAL-W2 version number; if 4 or greater, output is csv format!; otherwise, it's fixed width (8 space columns)}

\item{append.filename}{logical will append the named file with the new water balance file; if NULL, new.npt.filename will be used to write a new QDT file}

\item{maxElvDif_m}{numeric (meters) threshold in which to ignore water balance changes}
}
\value{
\item{fit}{a data.frame of the model to observed fit statistics}
\item{compElvs}{a data.frame of the daily elevations}
\item{compVol}{a data.frame of the daily volumes and flows}
}
\description{
This function will compare measured and modeled lake elevation output, then write a new inflow file to be run on the next W2 model run to help achieve simulated flows that match the measured flows. This version setup for River models
}
\examples{
waterBalance()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{balance}
\keyword{water}
