% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readW2ConInOut.r
\name{readW2ConInOut}
\alias{readW2ConInOut}
\title{Wrapper function for waterBalance function
This function will grab inputs from CE-QUAL-W2 model (w2_con.npt file), lookup filenames, and then use them as input to the waterBalance function within the w2r package.}
\usage{
readW2ConInOut(
  path,
  wd,
  seg,
  wb,
  RESSIMCode,
  elvVolCrvFl,
  write.files,
  maxElvDif_m
)
}
\arguments{
\item{path}{string # path to model}

\item{wd}{string # Generic path to find functions}

\item{seg}{numeric #Segment to look up in w2_con.npt}

\item{wb}{numeric #Water body to look up in w2_con.npt}

\item{RESSIMCode}{string #RES-SIM name for boundary conditions}

\item{elvVolCrvFl}{data.frame #Elevation - Volume Curve}

\item{write.files}{logical #T/F on whether or not to write file}

\item{maxElvDif_m}{numeric (meters) threshold in which to ignore water balance changes}
}
\value{
\item{watbal}{output from the waterBalance function}
}
\description{
Wrapper function for waterBalance function
This function will grab inputs from CE-QUAL-W2 model (w2_con.npt file), lookup filenames, and then use them as input to the waterBalance function within the w2r package.
}
\examples{
readW2ConInOut()
}
\author{
Norman Buccola
}
\keyword{CEQUALW2}
\keyword{QDT}
\keyword{balance}
\keyword{water}
